TALENTHUB-main/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Marshmallow>=0.14 # Using a common version, can be updated if needed
marshmallow-sqlalchemy>=0.23 # For SQLAlchemyAutoSchema
Flask-CORS>=3.0
gunicorn
honcho
pandas
openpyxl
